refactor for future endpoints
