+++ packages/uma/config/routes/policies.json
@@ -11,7 +11,7 @@
             ],
             "handler": {
                 "@type": "PolicyRequestHandler",
-                "store": {
+                "storage": {
                     "@id": "urn:uma:default:RulesStorage"
                 }
             },
